// frontend/src/lib/api/generated/api.ts
export interface TenantProfileResponseDto {
    'id'?: string;
    'tenantId'?: string;
    'fullName'?: string;
    'email'?: string;
    'phone'?: string;
    'identificationReference'?: string;
    'createdAt'?: string;
    'updatedAt'?: string;
}